Refactor: Ta dando tela, MAS TA TUDO QUEBRADO

+++ srcs/parser/map_grid_utils.c
@@ -6,7 +6,7 @@
 /*   By: lbarreto <lbarreto@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/30 16:20:58 by lbarreto          #+#    #+#             */
-/*   Updated: 2025/07/07 23:36:10 by lbarreto         ###   ########.fr       */
+/*   Updated: 2025/07/16 16:41:32 by lbarreto         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int	get_grid_x_size(char **grid)
 	bigger_line = 0;
 	while (grid[i])
 	{
-		if (ft_strlen(grid[i]) > bigger_line)
+		if ((int)ft_strlen(grid[i]) > bigger_line)
 			bigger_line = ft_strlen(grid[i]);
 		i++;
 	}
